evil tensors

+++ test/experiments/3D_12031.cu
@@ -11,7 +11,7 @@ using Conf0 = Config<Shape<100,80,60>,
 
                 using Conf2 = Config<Shape<100,80,60>,
                                                       Shape<10,8,6>,
-                                                      double, __half, float, __half,
+                                                      double, __half, float, float,
                                                       uint64_t>;
